feat(ui): one send screen, not two — same amounts, no prefilled money, nothing clipped
A review of the screens people use to move money, on web and mobile side by side.
Every change is a removal or an alignment; nothing new to learn.

Web
- The number field clipped its last digits at phone width (108px for a 115px
  value). It now has room for all nine.
- The amount arrived prefilled with 50 000 XAF. A prefilled sum on a money screen
  is a mistake waiting to be confirmed; it now starts empty, Continue stays off
  until a sum is typed, and the field shows a placeholder rather than a literal 0.
- Quick-pick amounts are now the same four on web and mobile (5 000 · 10 000 ·
  25 000 · 50 000), defined once in shared/domain.ts, shown as full numbers rather
  than "10k".
- The Delivered screen clipped the date ("08 September 2026 at 10…"); a shorter
  month keeps it on one line in both languages.
- Getting paid had no way in from the app shell — /receive was reachable only from
  the marketing pages. "Get paid" is now in the bottom nav.

Mobile
- Review showed the recipient's name and operator but not the number. The number
  is what a sender checks a payment against; it is now under the name.
- Receive said the same thing three times (title, subtitle, intro paragraph). The
  paragraph is gone; the fix-your-saved-number message stays when it applies.

Verified on the iOS simulator and the web app at 390px. Suite green (862 checks).

+++ app/src/pages/send/Success.tsx
@@ -11,7 +11,7 @@ function fullPhone(p: Payment): string {
   return COUNTRIES[p.recipient.country].dial + " " + p.recipient.phone;
 }
 function when(p: Payment, lang: "en" | "fr"): string {
-  return new Date(p.createdAt).toLocaleString(lang === "fr" ? "fr-FR" : "en-GB", { day: "2-digit", month: "long", year: "numeric", hour: "2-digit", minute: "2-digit" });
+  return new Date(p.createdAt).toLocaleString(lang === "fr" ? "fr-FR" : "en-GB", { day: "2-digit", month: "short", year: "numeric", hour: "2-digit", minute: "2-digit" });
 }
 
 export function Receipt({ payment, onClose }: { payment: Payment; onClose: () => void }) {
